+++ Apps/Examples/Examples/All Examples/Annotations/DynamicViewAnnotationExample.swift
@@ -225,7 +225,7 @@ private final class Route {
         guard !displayed, let mapView else { return }
         displayed = true
 
-        func colorExpression(normal: String, selected: String) -> MapboxMaps.Expression {
+        func colorExpression(normal: String, selected: String) -> Exp {
             Exp(.switchCase) {
                 Exp(.boolean) {
                     Exp(.featureState) { "selected" }