+++ Apps/Examples/Examples/All Examples/Annotations/PointAnnotationClusteringExample.swift
@@ -103,7 +103,7 @@ final class PointAnnotationClusteringExample: UIViewController, ExampleProtocol 
         }
 
         // Create a cluster property to add to each cluster
-        let clusterProperties: [String: MapboxMaps.Expression] = [
+        let clusterProperties: [String: Exp] = [
             "sum": sumExpression
         ]