+++ Apps/Examples/Examples/All Examples/Annotations/SymbolClusteringExample.swift
@@ -77,7 +77,7 @@ final class SymbolClusteringExample: UIViewController, ExampleProtocol {
         }
 
         // Add the expressions to the cluster as ClusterProperties so they can be accessed below
-        let clusterProperties: [String: MapboxMaps.Expression] = [
+        let clusterProperties: [String: Exp] = [
             "max": maxExpression,
             "in_e9": ine9Expression,
             "sum": sumExpression