+++ Sources/MapboxMaps/Foundation/Extensions/Core/RenderedQueryOptions.swift
@@ -7,7 +7,7 @@ extension RenderedQueryOptions {
     /// - Parameters:
     ///   - layerIds: List of layer identifiers to limit the feature query to
     ///   - filter: Filter to a type of feature with an expression
-    public convenience init(layerIds: [String]?, filter: Expression?) {
+    public convenience init(layerIds: [String]?, filter: Exp?) {
         var filterJson: Any?
         if let filter = filter {
             do {
@@ -20,15 +20,15 @@ extension RenderedQueryOptions {
         self.init(__layerIds: layerIds, filter: filterJson)
     }
 
-    /// Filters the returned features with an ``Expression``.
-    public var filter: Expression? {
+    /// Filters the returned features with an ``Exp``.
+    public var filter: Exp? {
 
         guard let filter = __filter else { return nil }
 
         var filterExp: Exp?
         do {
             let filterData = try JSONSerialization.data(withJSONObject: filter, options: [])
-            filterExp = try JSONDecoder().decode(Expression.self, from: filterData)
+            filterExp = try JSONDecoder().decode(Exp.self, from: filterData)
         } catch {
             Log.error(forMessage: "Filter expression could not be decoded", category: "RenderedQueryOptions")
         }